requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.0.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
cachetools=5.1.0=pypi_0
certifi=2022.12.7=pypi_0
charset-normalizer=2.1.1=pypi_0
cycler=0.11.0=pypi_0
expat=2.7.1=h6a678d5_0
fonttools=4.33.3=pypi_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.75.1=pypi_0
idna=3.4=pypi_0
imageio=2.19.2=pypi_0
imageio-ffmpeg=0.4.7=pypi_0
importlib-metadata=4.11.3=pypi_0
kiwisolver=1.4.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
markdown=3.3.7=pypi_0
markupsafe=3.0.3=pypi_0
matplotlib=3.5.2=pypi_0
ncurses=6.5=h7934f7d_0
numpy=1.22.4=pypi_0
oauthlib=3.2.0=pypi_0
opencv-python=4.5.5.64=pypi_0
openssl=3.0.17=h5eee18b_0
packaging=21.3=pypi_0
pillow=11.0.0=pypi_0
pip=25.2=pyhc872135_0
protobuf=3.20.1=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=pypi_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.8.2=pypi_0
readline=8.3=hc2a1206_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scipy=1.7.3=pypi_0
setuptools=80.9.0=py310h06a4308_0
six=1.17.0=pypi_0
some-package=0.1=pypi_0
sqlite=3.50.2=hb25bd0a_1
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tk=8.6.15=h54e0aa7_0
torch=1.11.0+cu113=pypi_0
torchvision=0.12.0+cu113=pypi_0
tqdm=4.64.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025b=h04d1e81_0
urllib3=1.26.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zipp=3.8.0=pypi_0
zlib=1.3.1=hb25bd0a_0
